# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TokMan API
  description: |
    TokMan is a token-aware CLI proxy that intercepts commands and applies
    a 31-layer compression pipeline to reduce token usage for AI coding assistants.
    
    ## Authentication
    Most endpoints require an API key via the `X-API-Key` header.
    
    ## Rate Limiting
    Rate limits are applied per API key. Check `X-RateLimit-Remaining` header.
  version: 1.0.0
  contact:
    name: TokMan Support
    url: https://github.com/GrayCodeAI/tokman
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.tokman.dev/v1
    description: Production server
...
paths:
  /compress:
    post:
      tags:
        - compress
      summary: Compress text content
      description: |
        Apply the 31-layer compression pipeline to reduce token count.
        Supports multiple presets (fast, balanced, full) and custom configurations.
      operationId: compressText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompressRequest'
      responses:
        '200':
          description: Successfully compressed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompressResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /analyze:
    post:
      tags:
        - analyze
      summary: Analyze text tokens
      description: |
        Get detailed token analysis including count, entropy,
        and compression potential estimation.
      operationId: analyzeText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
...
components:
  schemas:
    CompressRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Text content to compress
        preset:
          type: string
          enum: [fast, balanced, full]
          default: balanced
          description: Compression preset
        mode:
          type: string
          enum: [none, minimal, aggressive]
          default: minimal
          description: Compression mode
        budget:
          type: integer
          minimum: 100
          description: Token budget limit
        queryIntent:
          type: string
          description: Query context for goal-driven compression
        enableLLM:
          type: boolean
          default: false
          description: Enable LLM-based summarization
        sessionId:
          type: string
          format: uuid
          description: Optional session for context persistence
    CompressResponse:
      type: object
      properties:
        compressed:
          type: string
          description: Compressed text content
        originalTokens:
          type: integer
          description: Original token count
        compressedTokens:
          type: integer
          description: Compressed token count
        tokensSaved:
          type: integer
          description: Tokens saved by compression
        compressionRatio:
          type: number
          format: float
          description: Compression ratio (0.0-1.0)
        layersApplied:
          type: array
          items:
            type: string
          description: List of layers that were applied
        processingTimeMs:
          type: integer
          description: Processing time in milliseconds
...
    AnalyzeRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Text content to analyze
        detailed:
          type: boolean
          default: false
          description: Include detailed layer analysis
    AnalyzeResponse:
      type: object
      properties:
        tokenCount:
          type: integer
          description: Total token count
        lineCount:
          type: integer
          description: Number of lines
        charCount:
          type: integer
          description: Character count
        entropy:
          type: number
          format: float
          description: Content entropy score
        estimatedCompression:
          type: number
          format: float
          description: Estimated compression potential (0.0-1.0)
        language:
          type: string
          description: Detected programming language
        layerAnalysis:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/LayerAnalysis'
    LayerAnalysis:
      type: object
      properties:
        applicable:
          type: boolean
        estimatedSavings:
          type: integer
        priority:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []